requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pillow
pytesseract
PyMuPDF
openai
python-dotenv
